offer list, dedail

+++ archive/management/commands/load_offer.py
@@ -14,41 +14,50 @@ from catalog.models import *
 
 class Command(BaseCommand):
 
+    def add_arguments(self, parser):
+        
+        parser.add_argument(
+            '-f',
+            dest='force',
+            help='Очищает таблицу',
+        )
+
     def handle(self, *args, **options):
-        print 'Clearing the table'
-        OfferMessage.objects.all().delete()
-        Offer.objects.all().delete()
-        print 'Start loading..'
-        cnt = 0
-        for a in Announcement.objects.all().order_by('-id'):
-            cnt = cnt + 1
-            #if cnt == 10:
-            #    break
-            print 'Process announcement...%s' % a.id
-            for u in User.objects.exclude(id=a.user_id):
-                print 'Saving .....'
-                o = Offer()
-                o.user = u
-                o.message = 'test message'
-                o.url = 'test url'
-                o.price = 10
-                o.announcement = a
-                o.save()
+        if options['force'] == 1:
+            print 'Clearing the table'
+            OfferMessage.objects.all().delete()
+            Offer.objects.all().delete()
 
-                m = OfferMessage()
-                m.offer = o
-                m.new_price = 20
-                m.message = 'I have a new dial'
-                m.user = u
-                m.save()
+        cnt = Announcement.objects.all().count()
+        if cnt>0:
 
-                m = OfferMessage()
-                m.offer = o
-                m.new_price = 20
-                m.message = 'I have a new dial 2'
-                m.user = u
-                m.save()
+            print 'Start loading..'
+            cnt = 0
+            for a in Announcement.objects.all().order_by('-id'):
+                cnt = cnt + 1
+                #if cnt == 10:
+                #    break
+                print 'Process announcement...%s' % a.id
+                for u in User.objects.exclude(id=a.user_id):
+                    print 'Saving .....'
+                    o = Offer()
+                    o.seller = u
+                    o.buyer = a.user
+                    o.message = 'Тестовое предложение от продавца %s' % u 
+                    o.url = 'test url'
+                    o.price = 10
+                    o.announcement = a
+                    o.save()
 
-            #print a.user
-            #break
-           
+                    for i in range(0,3):
+                        m = OfferMessage()
+                        m.offer = o
+                        m.new_price = 20
+                        m.message = 'Тестовое сообщение другой цены от продавца %s' % u 
+                        m.user = u
+                        m.save()
+
+                  
+                #print a.user
+                #break
+
